refactor(step3): 클래스 통일
