+++ src/test/java/step3/game/PositiveNumberTest.java
@@ -1,9 +1,8 @@
-package step3;
+package step3.game;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.game.PositiveNumber;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
